requirements.txt
fastapi>=0.100.0
uvicorn>=0.20.0
pydantic>=2.0.0
openai>=1.0.0
langgraph>=0.0.10
langchain>=0.0.300
langchain-openai>=0.0.1
python-multipart>=0.0.6
httpx>=0.24.0
numpy>=1.21.0
pandas>=1.5.0
scikit-learn>=1.0.0
python-dotenv>=0.19.0
mlflow>=2.0.0
firebase-admin>=6.0.0
google-cloud-firestore>=2.0.0
google-cloud-bigquery>=3.0.0 